Generalized frequency dependency
