+++ include/tudat/astro/observation_models/angularPositionObservationModel.h
@@ -76,9 +76,11 @@ public:
             const LinkEndType linkEndAssociatedWithTime,
             std::vector< double >& linkEndTimes,
             std::vector< Eigen::Matrix< double, 6, 1 > >& linkEndStates,
-            const std::shared_ptr< ObservationAncilliarySimulationSettings > ancilliarySetings = nullptr )
-
+            const std::shared_ptr< ObservationAncilliarySimulationSettings > ancilliarySetingsInput = nullptr )
     {
+        std::shared_ptr< ObservationAncilliarySimulationSettings > ancilliarySetings;
+        this->setFrequencyProperties( time, linkEndAssociatedWithTime, lightTimeCalculator_, ancilliarySetingsInput, ancilliarySetings );
+
         // Check link end associated with input time and compute observable
         bool isTimeAtReception;
         if( linkEndAssociatedWithTime == receiver )
@@ -107,11 +109,6 @@ public:
         ObservationScalarType lightTime = lightTimeCalculator_->calculateLightTimeWithLinkEndsStates(
                 receiverState, transmitterState, time, isTimeAtReception, ancilliarySetings );
 
-        //        // Compute spherical relative position
-        //        Eigen::Matrix< ObservationScalarType, 3, 1 > sphericalRelativeCoordinates =
-        //                coordinate_conversions::convertCartesianToSpherical< ObservationScalarType >(
-        //                    transmitterState.segment( 0, 3 ) - receiverState.segment( 0, 3 ) ).
-        //                template cast< ObservationScalarType >( );
 
         Eigen::Matrix< ObservationScalarType, 3, 1 > relativePosition = transmitterState.segment( 0, 3 ) - receiverState.segment( 0, 3 );
