+++ include/tudat/astro/observation_models/corrections/lightTimeCorrection.h
@@ -13,7 +13,7 @@
 
 #include "tudat/basics/basicTypedefs.h"
 #include "tudat/astro/observation_models/linkTypeDefs.h"
-#include "tudat/astro/observation_models/observationModel.h"
+#include "tudat/astro/observation_models/observationAncillarySettings.h"
 
 namespace tudat
 {
