+++ include/tudat/astro/observation_models/dsnNWayRangeObservationModel.h
@@ -35,6 +35,10 @@ class DsnNWayRangeObservationModel : public ObservationModel< 1, ObservationScal
 {
 public:
     typedef Eigen::Matrix< ObservationScalarType, 6, 1 > StateType;
+    
+    using ObservationModel< 1, ObservationScalarType, TimeType >::turnaroundRatio_;
+    using ObservationModel< 1, ObservationScalarType, TimeType >::timeScaleConverter_;
+    using ObservationModel< 1, ObservationScalarType, TimeType >::frequencyInterpolator_;
 
     /*! Constructor.
      *
@@ -57,9 +61,11 @@ public:
                     std::map< LinkEndType, std::shared_ptr< ground_stations::GroundStationState > >( ) ):
         ObservationModel< 1, ObservationScalarType, TimeType >( dsn_n_way_range, linkEnds, observationBiasCalculator ),
         lightTimeCalculator_( lightTimeCalculator ), numberOfLinkEnds_( linkEnds.size( ) ),
-        transmittingFrequencyCalculator_( transmittingFrequencyCalculator ), turnaroundRatio_( turnaroundRatio ),
         stationStates_( groundStationStates )
     {
+        this->setFrequencyInterpolatorAndTurnaroundRatio(
+                transmittingFrequencyCalculator, turnaroundRatio );
+                
         if( !std::is_same< Time, TimeType >::value )
         {
             //            std::cerr<<
@@ -75,7 +81,6 @@ public:
                     "link ends, " +
                     std::to_string( numberOfLinkEnds_ ) + "were selected." );
         }
-        terrestrialTimeScaleConverter_ = earth_orientation::createDefaultTimeConverter( );
     }
 
     //! Destructor
@@ -170,8 +175,8 @@ public:
         if( lightTimeCalculator_->doCorrectionsNeedFrequency( ) )
         {
             setTransmissionReceptionFrequencies( lightTimeCalculator_,
-                                                 terrestrialTimeScaleConverter_,
-                                                 transmittingFrequencyCalculator_,
+                                                 timeScaleConverter_,
+                                                 frequencyInterpolator_,
                                                  time,
                                                  linkEndAssociatedWithTime,
                                                  ancillarySettings,
@@ -184,19 +189,19 @@ public:
         Eigen::Vector3d nominalReceivingStationState = ( stationStates_.count( receiver ) == 0 )
                 ? Eigen::Vector3d::Zero( )
                 : stationStates_.at( receiver )->getNominalCartesianPosition( );
-        TimeType utcReceptionTime = terrestrialTimeScaleConverter_->getCurrentTime< TimeType >(
+        TimeType utcReceptionTime = timeScaleConverter_->template getCurrentTime< TimeType >(
                 basic_astrodynamics::tdb_scale, basic_astrodynamics::utc_scale, time, nominalReceivingStationState );
-        TimeType utcTransmissionTime = terrestrialTimeScaleConverter_->getCurrentTime< TimeType >(
+        TimeType utcTransmissionTime = timeScaleConverter_->template getCurrentTime< TimeType >(
                 basic_astrodynamics::tdb_scale, basic_astrodynamics::utc_scale, time - lightTime, nominalReceivingStationState );
 
         ObservationScalarType uplinkFrequency =
-                transmittingFrequencyCalculator_->template getTemplatedCurrentFrequency< ObservationScalarType, TimeType >(
+                frequencyInterpolator_->template getTemplatedCurrentFrequency< ObservationScalarType, TimeType >(
                         utcTransmissionTime );
         ancillarySettings->setAncilliaryDoubleData( observation_models::range_conversion_factor,
                                                     physical_constants::SPEED_OF_LIGHT / ( uplinkFrequency * conversionFactor ) );
 
         ObservationScalarType transmitterFrequencyIntegral =
-                transmittingFrequencyCalculator_->template getTemplatedFrequencyIntegral< ObservationScalarType, TimeType >(
+                frequencyInterpolator_->template getTemplatedFrequencyIntegral< ObservationScalarType, TimeType >(
                         utcTransmissionTime, utcReceptionTime );
         ObservationScalarType rangeUnitIntegral = conversionFactor * transmitterFrequencyIntegral;
 
@@ -219,14 +224,6 @@ private:
 
     // Number of link ends
     unsigned int numberOfLinkEnds_;
-
-    // Object returning the transmitted frequency as the transmitting link end
-    std::shared_ptr< ground_stations::StationFrequencyInterpolator > transmittingFrequencyCalculator_;
-
-    // Function returning the turnaround ratio for given uplink and downlink bands
-    std::function< double( FrequencyBands uplinkBand, FrequencyBands downlinkBand ) > turnaroundRatio_;
-
-    std::shared_ptr< earth_orientation::TerrestrialTimeScaleConverter > terrestrialTimeScaleConverter_;
 
     std::map< LinkEndType, std::shared_ptr< ground_stations::GroundStationState > > stationStates_;
 };
